requirements.txt
feedparser>=6.0.0
praw>=7.7.0
anthropic>=0.40.0
python-telegram-bot>=21.0.0
APScheduler>=3.10.0
python-dotenv>=1.0.0
pytz>=2024.1
